[chore] Bump most crates to 2024 edition (#24040)
## Description 

- Bump all crates except `sui-graphql-rpc`, the lifetime errors in that
crate are a bit too involved for this PR

## Test plan 

- clippy

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

+++ crates/sui-data-ingestion-core/src/reader.rs
@@ -9,8 +9,8 @@ use futures::StreamExt;
 use mysten_metrics::spawn_monitored_task;
 #[cfg(not(target_os = "macos"))]
 use notify::{RecommendedWatcher, RecursiveMode};
-use object_store::path::Path;
 use object_store::ObjectStore;
+use object_store::path::Path;
 use std::ffi::OsString;
 use std::fs;
 use std::path::PathBuf;
@@ -313,10 +313,10 @@ impl CheckpointReader {
         for entry in fs::read_dir(self.path.clone())? {
             let entry = entry?;
             let filename = entry.file_name();
-            if let Some(sequence_number) = Self::checkpoint_number_from_file_path(&filename) {
-                if sequence_number < watermark {
-                    fs::remove_file(entry.path())?;
-                }
+            if let Some(sequence_number) = Self::checkpoint_number_from_file_path(&filename)
+                && sequence_number < watermark
+            {
+                fs::remove_file(entry.path())?;
             }
         }
         Ok(())
